+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking changes
 
+* Change stream.* fields to dataset.* fields. [#](https://github.com/elastic/package-registry/pull/)
+
 ### Bugfixes
 
 ### Added